+++ t-rex-gdal/src/gdal_ds.rs
@@ -13,7 +13,7 @@ use core::Config;
 use datasource::DatasourceInput;
 use gdal;
 use gdal::spatial_ref::{CoordTransform, SpatialRef};
-use gdal::vector::{Dataset, FieldValue, Geometry, OGRwkbGeometryType};
+use gdal::vector::{Dataset, FieldValue, Geometry};
 use gdal_sys;
 use std::collections::BTreeMap;
 use std::path::Path;
@@ -60,6 +60,166 @@ fn geom_type_name(ogr_type: OGRwkbGeometryType::Type) -> Option<String> {
     }
 }
 
+// OGRwkbGeometryType from GDAL 2.2 bindings
+#[allow(non_snake_case, non_upper_case_globals)]
+pub mod OGRwkbGeometryType {
+    /// List of well known binary geometry types.  These are used within the BLOBs
+    /// but are also returned from OGRGeometry::getGeometryType() to identify the
+    /// type of a geometry object.
+    pub type Type = u32;
+    /// < unknown type, non-standard
+    pub const wkbUnknown: Type = 0;
+    /// < 0-dimensional geometric object, standard WKB
+    pub const wkbPoint: Type = 1;
+    /// < 1-dimensional geometric object with linear
+    /// interpolation between Points, standard WKB
+    pub const wkbLineString: Type = 2;
+    /// < planar 2-dimensional geometric object defined
+    /// by 1 exterior boundary and 0 or more interior
+    /// boundaries, standard WKB
+    pub const wkbPolygon: Type = 3;
+    /// < GeometryCollection of Points, standard WKB
+    pub const wkbMultiPoint: Type = 4;
+    /// < GeometryCollection of LineStrings, standard WKB
+    pub const wkbMultiLineString: Type = 5;
+    /// < GeometryCollection of Polygons, standard WKB
+    pub const wkbMultiPolygon: Type = 6;
+    /// < geometric object that is a collection of 1
+    /// or more geometric objects, standard WKB
+    pub const wkbGeometryCollection: Type = 7;
+    /// < one or more circular arc segments connected end to end,
+    /// ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbCircularString: Type = 8;
+    /// < sequence of contiguous curves, ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbCompoundCurve: Type = 9;
+    /// < planar surface, defined by 1 exterior boundary
+    /// and zero or more interior boundaries, that are curves.
+    /// ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbCurvePolygon: Type = 10;
+    /// < GeometryCollection of Curves, ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbMultiCurve: Type = 11;
+    /// < GeometryCollection of Surfaces, ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbMultiSurface: Type = 12;
+    /// < Curve (abstract type). ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCurve: Type = 13;
+    /// < Surface (abstract type). ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbSurface: Type = 14;
+    /// < a contiguous collection of polygons, which share common boundary segments,
+    /// ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbPolyhedralSurface: Type = 15;
+    /// < a PolyhedralSurface consisting only of Triangle patches
+    /// ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTIN: Type = 16;
+    /// < a Triangle. ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTriangle: Type = 17;
+    /// < non-standard, for pure attribute records
+    pub const wkbNone: Type = 100;
+    /// < non-standard, just for createGeometry()
+    pub const wkbLinearRing: Type = 101;
+    /// < wkbCircularString with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbCircularStringZ: Type = 1008;
+    /// < wkbCompoundCurve with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbCompoundCurveZ: Type = 1009;
+    /// < wkbCurvePolygon with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbCurvePolygonZ: Type = 1010;
+    /// < wkbMultiCurve with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbMultiCurveZ: Type = 1011;
+    /// < wkbMultiSurface with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.0
+    pub const wkbMultiSurfaceZ: Type = 1012;
+    /// < wkbCurve with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCurveZ: Type = 1013;
+    /// < wkbSurface with Z component. ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbSurfaceZ: Type = 1014;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbPolyhedralSurfaceZ: Type = 1015;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTINZ: Type = 1016;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTriangleZ: Type = 1017;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbPointM: Type = 2001;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbLineStringM: Type = 2002;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbPolygonM: Type = 2003;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiPointM: Type = 2004;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiLineStringM: Type = 2005;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiPolygonM: Type = 2006;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbGeometryCollectionM: Type = 2007;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCircularStringM: Type = 2008;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCompoundCurveM: Type = 2009;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCurvePolygonM: Type = 2010;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiCurveM: Type = 2011;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiSurfaceM: Type = 2012;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCurveM: Type = 2013;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbSurfaceM: Type = 2014;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbPolyhedralSurfaceM: Type = 2015;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTINM: Type = 2016;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTriangleM: Type = 2017;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbPointZM: Type = 3001;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbLineStringZM: Type = 3002;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbPolygonZM: Type = 3003;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiPointZM: Type = 3004;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiLineStringZM: Type = 3005;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiPolygonZM: Type = 3006;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbGeometryCollectionZM: Type = 3007;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCircularStringZM: Type = 3008;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCompoundCurveZM: Type = 3009;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCurvePolygonZM: Type = 3010;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiCurveZM: Type = 3011;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbMultiSurfaceZM: Type = 3012;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbCurveZM: Type = 3013;
+    /// < ISO SQL/MM Part 3. GDAL &gt;= 2.1
+    pub const wkbSurfaceZM: Type = 3014;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbPolyhedralSurfaceZM: Type = 3015;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTINZM: Type = 3016;
+    /// < ISO SQL/MM Part 3. Reserved in GDAL &gt;= 2.1 but not yet implemented
+    pub const wkbTriangleZM: Type = 3017;
+    /// < 2.5D extension as per 99-402
+    pub const wkbPoint25D: Type = 2147483649;
+    /// < 2.5D extension as per 99-402
+    pub const wkbLineString25D: Type = 2147483650;
+    /// < 2.5D extension as per 99-402
+    pub const wkbPolygon25D: Type = 2147483651;
+    /// < 2.5D extension as per 99-402
+    pub const wkbMultiPoint25D: Type = 2147483652;
+    /// < 2.5D extension as per 99-402
+    pub const wkbMultiLineString25D: Type = 2147483653;
+    /// < 2.5D extension as per 99-402
+    pub const wkbMultiPolygon25D: Type = 2147483654;
+    /// < 2.5D extension as per 99-402
+    pub const wkbGeometryCollection25D: Type = 2147483655;
+}
+
 trait ToGeo {
     fn to_geo(&self, srid: Option<i32>) -> GeometryType;
 }
@@ -78,8 +238,10 @@ impl ToGeo for Geometry {
         };
 
         match geometry_type {
-            OGRwkbGeometryType::wkbPoint | OGRwkbGeometryType::wkbPoint25D => {
-                // GDAL 2.1: | OGRwkbGeometryType::wkbPointM | OGRwkbGeometryType::wkbPointZM
+            OGRwkbGeometryType::wkbPoint
+            | OGRwkbGeometryType::wkbPoint25D
+            | OGRwkbGeometryType::wkbPointM
+            | OGRwkbGeometryType::wkbPointZM => {
                 let (x, y, _) = self.get_point(0);
                 GeometryType::Point(geom::Point {
                     x: x,
@@ -87,7 +249,10 @@ impl ToGeo for Geometry {
                     srid: srid,
                 })
             }
-            OGRwkbGeometryType::wkbMultiPoint | OGRwkbGeometryType::wkbMultiPoint25D => {
+            OGRwkbGeometryType::wkbMultiPoint
+            | OGRwkbGeometryType::wkbMultiPoint25D
+            | OGRwkbGeometryType::wkbMultiPointM
+            | OGRwkbGeometryType::wkbMultiPointZM => {
                 let point_count = self.geometry_count();
                 let coords = (0..point_count)
                     .map(|n| match unsafe { self._get_geometry(n) }.to_geo(srid) {
@@ -99,7 +264,10 @@ impl ToGeo for Geometry {
                     srid: srid,
                 })
             }
-            OGRwkbGeometryType::wkbLineString => {
+            OGRwkbGeometryType::wkbLineString
+            | OGRwkbGeometryType::wkbLineString25D
+            | OGRwkbGeometryType::wkbLineStringM
+            | OGRwkbGeometryType::wkbLineStringZM => {
                 let coords = self
                     .get_point_vec()
                     .iter()
@@ -113,7 +281,10 @@ impl ToGeo for Geometry {
                     srid: srid,
                 })
             }
-            OGRwkbGeometryType::wkbMultiLineString | OGRwkbGeometryType::wkbMultiLineString25D => {
+            OGRwkbGeometryType::wkbMultiLineString
+            | OGRwkbGeometryType::wkbMultiLineString25D
+            | OGRwkbGeometryType::wkbMultiLineStringM
+            | OGRwkbGeometryType::wkbMultiLineStringZM => {
                 let string_count = self.geometry_count();
                 let strings = (0..string_count)
                     .map(|n| match unsafe { self._get_geometry(n) }.to_geo(srid) {
@@ -125,7 +296,10 @@ impl ToGeo for Geometry {
                     srid: srid,
                 })
             }
-            OGRwkbGeometryType::wkbPolygon | OGRwkbGeometryType::wkbPolygon25D => {
+            OGRwkbGeometryType::wkbPolygon
+            | OGRwkbGeometryType::wkbPolygon25D
+            | OGRwkbGeometryType::wkbPolygonM
+            | OGRwkbGeometryType::wkbPolygonZM => {
                 let ring_count = self.geometry_count();
                 let rings = (0..ring_count).map(|n| ring(n)).collect();
                 GeometryType::Polygon(geom::Polygon {
@@ -133,7 +307,10 @@ impl ToGeo for Geometry {
                     srid: srid,
                 })
             }
-            OGRwkbGeometryType::wkbMultiPolygon | OGRwkbGeometryType::wkbMultiPolygon25D => {
+            OGRwkbGeometryType::wkbMultiPolygon
+            | OGRwkbGeometryType::wkbMultiPolygon25D
+            | OGRwkbGeometryType::wkbMultiPolygonM
+            | OGRwkbGeometryType::wkbMultiPolygonZM => {
                 let string_count = self.geometry_count();
                 let strings = (0..string_count)
                     .map(|n| match unsafe { self._get_geometry(n) }.to_geo(srid) {
@@ -156,7 +333,11 @@ impl ToGeo for Geometry {
                                                  })
             }
             */
-            geom_type => panic!("Unsupported geometry type {}", &ogr_type_name(geom_type)),
+            geom_type => panic!(
+                "Unsupported geometry type {} ({})",
+                geom_type,
+                &ogr_type_name(geom_type)
+            ),
         }
     }
 }